Move build_file_list into common module
