+++ rust/datafusion/src/execution/physical_plan/common.rs
@@ -17,9 +17,12 @@
 
 //! Defines common code used in execution plans
 
+use std::fs;
+use std::fs::metadata;
+use std::fs::File;
 use std::sync::{Arc, Mutex};
 
-use crate::error::Result;
+use crate::error::{ExecutionError, Result};
 use crate::execution::physical_plan::BatchIterator;
 
 use arrow::datatypes::Schema;
@@ -74,4 +77,31 @@ pub fn collect(it: Arc<Mutex<dyn BatchIterator>>) -> Result<Vec<RecordBatch>> {
             Err(e) => return Err(e),
         }
     }
+}
+
+/// Recursively build a list of files in a directory with a given extension
+pub fn build_file_list(dir: &str, filenames: &mut Vec<String>, ext: &str) -> Result<()> {
+    let metadata = metadata(dir)?;
+    if metadata.is_file() {
+        if dir.ends_with(ext) {
+            filenames.push(dir.to_string());
+        }
+    } else {
+        for entry in fs::read_dir(dir)? {
+            let entry = entry?;
+            let path = entry.path();
+            if let Some(path_name) = path.to_str() {
+                if path.is_dir() {
+                    build_file_list(path_name, filenames, ext)?;
+                } else {
+                    if path_name.ends_with(ext) {
+                        filenames.push(path_name.to_string());
+                    }
+                }
+            } else {
+                return Err(ExecutionError::General("Invalid path".to_string()));
+            }
+        }
+    }
+    Ok(())
 }
